requirements.txt
requests>=2.28.0
tqdm>=4.65.0
colorama>=0.4.6
customtkinter>=5.2.0
plyer>=2.1.0
yt-dlp>=2026.3.0
ffmpeg-python>=0.2.0
eel>=0.16.0
pystray>=0.19.0
Pillow>=10.0.0
